+++ __tests__/components/Image.js
@@ -2,42 +2,15 @@ import Image from '../../src/components/Image';
 import React from 'react';
 import renderer from 'react-test-renderer';
 
-const mockLozad = {
-  observe: jest.fn()
-};
-
-jest.mock('../../src/lozad', () => {
-  return jest.fn().mockImplementation((el, options) => {
-    mockLozad.options = options;
-    return mockLozad;
-  });
-});
-
-const testRendererOptions = {
-  createNodeMock: () => {
-    return {
-      getBoundingClientRect: () => {
-        return {};
-      }
-    };
-  }
-};
-
 const testConfig = {
   domain: 'https://test.com',
   apiKey: 'abc123'
 };
 
 describe('Image', () => {
-  beforeEach(() => {
-    mockLozad.options = null;
-    mockLozad.observe.mockClear();
-  });
-
   it('should not crash when dont have expected parameters', () => {
     renderer.create(
-      <Image src={undefined} config={undefined} />,
-      testRendererOptions
+      <Image src={undefined} config={undefined} />
     );
   });
 
@@ -45,8 +18,7 @@ describe('Image', () => {
     expect(
       renderer
         .create(
-          <Image src={'https://image.here.com/logo.png'} config={testConfig} />,
-          testRendererOptions
+          <Image src={'https://image.here.com/logo.png'} config={testConfig} />
         )
         .toJSON()
     ).toMatchSnapshot();
@@ -59,8 +31,7 @@ describe('Image', () => {
           <Image
             src={'https://image.here.com/logo.png'}
             config={{ domain: 'test.com', apiKey: 'abc123' }}
-          />,
-          testRendererOptions
+          />
         )
         .toJSON()
     ).toMatchSnapshot();
@@ -74,8 +45,7 @@ describe('Image', () => {
             src={'https://image.here.com/logo.png'}
             config={testConfig}
             alt={'yo!'}
-          />,
-          testRendererOptions
+          />
         )
         .toJSON()
     ).toMatchSnapshot();
@@ -89,8 +59,7 @@ describe('Image', () => {
             src={'https://image.here.com/logo.png'}
             config={testConfig}
             lazy={false}
-          />,
-          testRendererOptions
+          />
         )
         .toJSON()
     ).toMatchSnapshot();
@@ -104,8 +73,7 @@ describe('Image', () => {
             src={'https://image.here.com/logo.png'}
             config={testConfig}
             op={'fit?size=100x200'}
-          />,
-          testRendererOptions
+          />
         )
         .toJSON()
     ).toMatchSnapshot();
@@ -119,8 +87,7 @@ describe('Image', () => {
             src={'data:ABCDEF'}
             config={testConfig}
             op={'fit?size=100x200'}
-          />,
-          testRendererOptions
+          />
         )
         .toJSON()
     ).toMatchSnapshot();
@@ -134,8 +101,7 @@ describe('Image', () => {
             src={'//image.here.com/logo.png'}
             config={testConfig}
             op={'fit?size=100x200'}
-          />,
-          testRendererOptions
+          />
         )
         .toJSON()
     ).toMatchSnapshot();
@@ -149,47 +115,9 @@ describe('Image', () => {
             src={'//image.here.com/logo.png?param=1'}
             config={testConfig}
             op={'fit?size=100x200'}
-          />,
-          testRendererOptions
+          />
         )
         .toJSON()
     ).toMatchSnapshot();
-  });
-
-  it('should use default config in intersection observer config', () => {
-    renderer
-      .create(
-        <Image src={'https://image.here.com/logo.png'} config={testConfig} lazy={true}/>,
-        testRendererOptions
-      );
-
-    expect(mockLozad.options).toEqual({
-      threshold: 0.01,
-      rootMargin: '40px 0px 0px 0px'
-    });
-  });
-
-  it('should override default values in intersection observer config', () => {
-    const config = Object.assign(
-      {},
-      testConfig,
-      {
-        lozadOptions: {
-          rootMargin: '10px 10px 10px 10px'
-        }
-      });
-
-    renderer
-      .create(
-        <Image src={'https://image.here.com/logo.png'}
-          config={config}
-          lazy={true}/>,
-        testRendererOptions
-      );
-
-    expect(mockLozad.options).toEqual({
-      threshold: 0.01,
-      rootMargin: '10px 10px 10px 10px'
-    });
   });
 });